fix: provide fallback static URL function
The function GetStaticURL uses a
the function `urls.JoinPath` introduced in hugo `v0.112.0`. Certain CMS applications such as CloudCannon use an embedded hugo binary to update a site in real time. The embedded binary cannot be updated to a compatible version at will. This fix introduces a fallback function for such scenarios.